# swagger.yaml
openapi: "3.0.3"
info:
  title: MySlotMate API
  description: |
    Backend API for **MySlotMate** – a slot-booking platform with wallet-based payments,
    Razorpay payouts, Aadhar (Setu OKYC) verification, AWS S3 file uploads,
    and real-time inbox messaging.

    Every response follows the standard envelope:
    ```json
    { "success": true|false, "data": ..., "message": "...", "error": "..." }
    ```
  version: "2.0.0"
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: "8080"
...
paths:
...
  /reviews/:
    post:
      tags: [Reviews]
      summary: Submit a review with rating & optional photos
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_Review"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /reviews/event/{eventID}:
    get:
      tags: [Reviews]
      summary: List reviews for an event
      operationId: getReviewsByEvent
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Array of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_ReviewList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /reviews/event/{eventID}/rating:
    get:
      tags: [Reviews]
      summary: Get aggregate rating for an event
      operationId: getAverageRating
      parameters:
        - $ref: "#/components/parameters/EventIDPath"
      responses:
        "200":
          description: Average rating
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      avg_rating:
                        type: number
                        format: double
                      total_reviews:
                        type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Inbox ───────────────────────────────────
...
  /support/:
    post:
      tags: [Support]
      summary: Create a support ticket (JSON or multipart with evidence)
      operationId: createSupportTicket
      description: |
        Accepts both `application/json` and `multipart/form-data`.
        For multipart, use form fields: `user_id`, `category`, `subject`, `message`,
        `reported_user_id`, `event_id`, `session_date`, `report_reason`, `is_urgent`,
        and file field `evidence` (multiple files allowed, uploaded to S3).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketRequest"
          multipart/form-data:
            schema:
              type: object
              required: [user_id, category, subject, message]
              properties:
                user_id:
                  type: string
                  format: uuid
                category:
                  $ref: "#/components/schemas/SupportCategory"
                subject:
                  type: string
                message:
                  type: string
                reported_user_id:
                  type: string
                  format: uuid
                event_id:
                  type: string
                  format: uuid
                session_date:
                  type: string
                  format: date-time
                report_reason:
                  $ref: "#/components/schemas/ReportReason"
                is_urgent:
                  type: string
                  enum: ["true", "false"]
                evidence:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SupportTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /support/{ticketID}:
    get:
      tags: [Support]
      summary: Get a support ticket by ID
      operationId: getSupportTicket
      parameters:
        - $ref: "#/components/parameters/TicketIDPath"
      responses:
        "200":
          description: Support ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SupportTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /support/user/{userID}:
    get:
      tags: [Support]
      summary: List all tickets for a user
      operationId: getUserTickets
      parameters:
        - $ref: "#/components/parameters/UserIDPath"
      responses:
        "200":
          description: Array of support tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SupportTicketList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /support/{ticketID}/message:
    post:
      tags: [Support]
      summary: Add a message to a ticket thread
      operationId: addSupportMessage
      parameters:
        - $ref: "#/components/parameters/TicketIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Message added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SupportTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /support/{ticketID}/resolve:
    post:
      tags: [Support]
      summary: Mark ticket as resolved
      operationId: resolveSupportTicket
      parameters:
        - $ref: "#/components/parameters/TicketIDPath"
      responses:
        "200":
          description: Ticket resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse_SupportTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Upload ──────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID token passed as `Authorization: Bearer <token>`. Admin endpoints additionally require the token's email to match the `ADMIN_EMAIL` env variable."

  # ── Parameters ──────────────────────────────
  parameters:
...
    EventIDPath:
      name: eventID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event UUID
    UserIDPath:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User UUID
...
    TicketIDPath:
      name: ticketID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Support ticket UUID
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string

    # ── Request Bodies ────────────────────────
...
    CreateReviewRequest:
      type: object
      required: [user_id, event_id, rating, description]
      properties:
        user_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        name:
          type: string
          nullable: true
          description: Optional reviewer display name
        description:
          type: string
          example: "Great event, well organised!"
        photo_urls:
          type: array
          items:
            type: string
          description: Pre-uploaded photo URLs from /upload
...
    CreateSupportTicketRequest:
      type: object
      required: [user_id, category, subject, message]
      properties:
        user_id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/SupportCategory"
        subject:
          type: string
        message:
          type: string
        reported_user_id:
          type: string
          format: uuid
          nullable: true
        event_id:
          type: string
          format: uuid
          nullable: true
        session_date:
          type: string
          format: date-time
          nullable: true
        report_reason:
          $ref: "#/components/schemas/ReportReason"
        evidence_urls:
          type: array
          items:
            type: string
          description: Pre-uploaded evidence URLs
        is_urgent:
          type: boolean
          default: false
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        name:
          type: string
          nullable: true
        description:
          type: string
        photo_urls:
          type: array
          items:
            type: string
        reply:
          type: array
          items:
            type: string
        sentiment_score:
          type: number
          format: double
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SupportTicket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/SupportCategory"
        reported_user_id:
          type: string
          format: uuid
          nullable: true
        subject:
          type: string
        messages:
          type: array
          items:
            type: object
            properties:
              sender:
                type: string
              text:
                type: string
              created_at:
                type: string
                format: date-time
        status:
          $ref: "#/components/schemas/SupportTicketStatus"
        event_id:
          type: string
          format: uuid
          nullable: true
        session_date:
          type: string
          format: date
          nullable: true
        report_reason:
          $ref: "#/components/schemas/ReportReason"
        evidence_urls:
          type: array
          items:
            type: string
        is_urgent:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SupportCategory:
      type: string
      enum: [report_participant, technical_support, policy_help]
    SupportTicketStatus:
      type: string
      enum: [open, in_progress, resolved, closed]
    ReportReason:
      type: string
      enum: [verbal_harassment, safety_concern, inappropriate_behavior, spam_or_scam]
...
    SuccessResponse_Review:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Review"
    SuccessResponse_ReviewList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
...
    SuccessResponse_SupportTicket:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/SupportTicket"
    SuccessResponse_SupportTicketList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
...